SHARE-384 Refactor copy paste errors (Admin)
ApproveProgramsAdmin.php class cant inherit from AllProgramsAdmin.php to get rid of CP error so I implemented a trait
